first test, testing schema output
